Minor corrections to two pieces
1. Is energy a component of momentum?

Had said that timelike vectors have positive length and spacelike have negative length. This was not correct; it is the square of the magnitude, ds^2 that is positive for timelike and negative for spacelike, so timelike is real and spacelike is imaginary.

2. What fraction of a kg of potatoes is particle rest mass?

Had stated that some small fraction of the mass would come from electromagnetic, weak and gravitational interactions. Not clear to what extent there is any positive (as opposed to negative) contribution from these, so removing as negligible either way.